Can now edit the app.js file for the app.
