--- WorkoutPrompter/components/StyledText.tsx
@@ -1,5 +1,0 @@
-import { Text, TextProps } from './Themed';
-
-export function MonoText(props: TextProps) {
-  return <Text {...props} style={[props.style, { fontFamily: 'SpaceMono' }]} />;
-}
